refactor: extract loader utilities into directory
