+++ loader/index.js
@@ -5,35 +5,9 @@
 const originalFetch = global.fetch;
 delete global.fetch;
 
-const { SourceMapConsumer, SourceMapGenerator, SourceNode } = require('source-map');
+const { SourceMapConsumer, SourceNode } = require('source-map');
 const { Template } = require('webpack');
-
-/**
- * Generates an identity source map from a source file.
- * @param {string} source The content of the source file.
- * @param {string} resourcePath The name of the source file.
- * @returns {import('source-map').RawSourceMap} The identity source map.
- */
-function getIdentitySourceMap(source, resourcePath) {
-  const sourceMap = new SourceMapGenerator();
-  sourceMap.setSourceContent(resourcePath, source);
-
-  source.split('\n').forEach((line, index) => {
-    sourceMap.addMapping({
-      source: resourcePath,
-      original: {
-        line: index + 1,
-        column: 0,
-      },
-      generated: {
-        line: index + 1,
-        column: 0,
-      },
-    });
-  });
-
-  return sourceMap.toJSON();
-}
+const getIdentitySourceMap = require('./utils/getIdentitySourceMap');
 
 /**
  * Gets a runtime template from provided function.
